Collection: rules cross the boundary, not SQL -- collection.c writes none
common/collection.c 43 -> 0. With phase 1 this takes the file from 95 to zero and
finishes the design chosen for it: the database module receives RULES and composes
every query from them.

dt_collection_rule_t -- property, mode, text, recursive -- is now what crosses.
dt_collection_t carries an array of them instead of `where_ext`, and
dt_collection_set_extended_where()/dt_collection_get_extended_where() are gone.
common/collection.c reads the per-rule conf keys into that array and passes it;
database/collection_query.c turns each rule into a WHERE, joins them with their
conjunctions, and builds the query. get_query_string() -- 528 lines and, usefully,
zero conf reads -- moved with it.

Four things the composer was doing that a layer-1 module must not, each inverted
rather than dragged along:

- resolving a LOCALISED module-order name to its stored version id. The caller
  installs a resolver (dt_collection_query_set_order_resolver); translation is
  presentation.
- listing those same names, which the value-list query interpolates into its SQL.
  The caller supplies the array (dt_collection_query_set_order_names); the module
  needs the strings, not the ability to translate.
- deciding whether a metadata field is hidden, and how film rolls are ordered.
  Both were conf reads inside the query builder. They are display preferences, so
  they arrive resolved, in dt_collection_values_request_t alongside the property.

Keeping develop/iop_order.h out is not cosmetic: including it took the module to 2
higher-layer includes against a boundary of 1, and the gate refused it.

Also moved: dt_collection_get_makermodels (pure SQL), get_images_for_rule (already
took a rule in argument form), get_property_values, and update_query's pair of
"find the first collected image not in this list" queries, which built their IN
list from a GList by hand -- the same shape as _num_set_predicate() in
style_repository.c, now built inside the module.

VERIFIED by output, not by reading. A moved composer cannot be checked with a
query-string diff, so the composed SQL itself was captured before and after
through two corpora, master-staged binary against HEAD-staged:

  90 combinations driving the RULES (property, mode, text, recursive, one and two
     rules, across filter flags and sort) -> 62 distinct queries, all identical;
  80 combinations driving the PARAMS (filter flags, sort, direction, text filter)
     -> 80 distinct queries, all identical.

Ratchet: 58 -> 52 handles, 24 -> 22 files.
